Shtuffz

+++ commands/sudo/gtfo.js
@@ -4,6 +4,7 @@ module.exports = {
 	useage: '',
 	isCommand: true,
 	execute(message, args) {
+		if(!global.good.includes(message.author.id)) return;
         global.client.destroy();
         process.exit(0);
 	},
